# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentFinding": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "output": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "state": {
            "$ref": "#/components/schemas/AgentState"
          }
        },
        "required": [
          "state"
        ],
        "title": "AgentFinding",
        "type": "object"
      },
      "AgentState": {
        "enum": [
          "AVAILABLE",
          "TIMEOUT",
          "INVALID_OUTPUT",
          "MODEL_ERROR",
          "DEPENDENCY_UNAVAILABLE"
        ],
        "title": "AgentState",
        "type": "string"
      },
...
      "DecisionListItem": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "environment": {
            "title": "Environment",
            "type": "string"
          },
          "escalation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "usage_units": {
            "title": "Usage Units",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "session_id",
          "state",
          "reason",
          "environment",
          "escalation_id",
          "usage_units",
          "created_at"
        ],
        "title": "DecisionListItem",
        "type": "object"
      },
      "DecisionState": {
        "enum": [
          "APPROVED",
          "APPROVED_WITH_REWRITE",
          "NEEDS_CLARIFICATION",
          "BLOCKED",
          "ESCALATED",
          "SYSTEM_UNAVAILABLE"
        ],
        "title": "DecisionState",
        "type": "string"
      },
      "DecisionTrace": {
        "additionalProperties": false,
        "properties": {
          "agent_findings": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AgentFinding"
            },
            "title": "Agent Findings",
            "type": "object"
          },
          "knowledge_release_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Release Id"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "profile_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Version Id"
          }
        },
        "required": [
          "agent_findings",
          "profile_version_id",
          "knowledge_release_id",
          "model_id"
        ],
        "title": "DecisionTrace",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublicDecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "end_user_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End User Id"
          },
          "input": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Input",
            "type": "string"
          },
          "metadata": {
            "title": "Metadata",
            "type": "object"
          },
          "session_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": [
          "input"
        ],
        "title": "PublicDecisionRequest",
        "type": "object"
      },
      "PublicDecisionResponse": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/SourceRecord"
            },
            "title": "Citations",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision_reason": {
            "title": "Decision Reason",
            "type": "string"
          },
          "decision_state": {
            "$ref": "#/components/schemas/DecisionState"
          },
          "escalation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "trace": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionTrace"
              },
              {
                "type": "null"
              }
            ]
          },
          "usage_units": {
            "title": "Usage Units",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "session_id",
          "answer",
          "decision_state",
          "decision_reason",
          "citations",
          "trace",
          "escalation_id",
          "usage_units",
          "created_at"
        ],
        "title": "PublicDecisionResponse",
        "type": "object"
      },
...
      "SourceRecord": {
        "additionalProperties": false,
        "properties": {
          "chunk_id": {
            "minLength": 1,
            "title": "Chunk Id",
            "type": "string"
          },
          "distance_or_similarity": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Distance Or Similarity",
            "type": "number"
          },
          "document_id": {
            "minLength": 1,
            "title": "Document Id",
            "type": "string"
          },
          "metadata": {
            "title": "Metadata",
            "type": "object"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "chunk_id",
          "text",
          "distance_or_similarity"
        ],
        "title": "SourceRecord",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AgentMesh Risk-Aware Support API",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/decisions": {
      "post": {
        "operationId": "create_decision_api_v1_decisions_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Decision",
        "tags": [
          "decisions"
        ]
      }
    },
    "/api/v1/decisions/{decision_id}": {
      "get": {
        "operationId": "get_public_decision_api_v1_decisions__decision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "title": "Decision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Public Decision",
        "tags": [
          "decisions"
        ]
      }
    },
...
    "/api/v1/organizations/{organization_id}/workspaces/{workspace_id}/decisions": {
      "get": {
        "operationId": "list_dashboard_decisions_api_v1_organizations__organization_id__workspaces__workspace_id__decisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User"
            }
          },
          {
            "in": "header",
            "name": "x-dev-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Email"
            }
          },
          {
            "in": "header",
            "name": "x-dev-aal",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Aal"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DecisionListItem"
                  },
                  "title": "Response List Dashboard Decisions Api V1 Organizations  Organization Id  Workspaces  Workspace Id  Decisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Dashboard Decisions",
        "tags": [
          "decisions"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/workspaces/{workspace_id}/decisions/{decision_id}": {
      "get": {
        "operationId": "get_dashboard_decision_api_v1_organizations__organization_id__workspaces__workspace_id__decisions__decision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "title": "Decision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User"
            }
          },
          {
            "in": "header",
            "name": "x-dev-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Email"
            }
          },
          {
            "in": "header",
            "name": "x-dev-aal",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Aal"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dashboard Decision",
        "tags": [
          "decisions"
        ]
      }
    },
...
  }
}